Added the branch: single file version
