--- data/__init__.py
@@ -1,1 +1,0 @@
-__author__ = 'justinarmstrong'
