--- data/components/collider.py
@@ -1,17 +1,0 @@
-__author__ = 'justinarmstrong'
-
-import pygame as pg
-from .. import constants as c
-
-class Collider(pg.sprite.Sprite):
-    """Invisible sprites placed overtop background parts
-    that can be collided with (pipes, steps, ground, etc."""
-    def __init__(self, x, y, width, height, name='collider'):
-        pg.sprite.Sprite.__init__(self)
-        self.image = pg.Surface((width, height)).convert()
-        #self.image.fill(c.RED)
-        self.rect = self.image.get_rect()
-        self.rect.x = x
-        self.rect.y = y
-        self.state = None
-
